Split part of app.js to init/components.js, init/filters.js, init/translations.js
