+++ src/init/components.js
@@ -1,0 +1,6 @@
+import Vue from 'vue'
+import ProgressLinearColor from '../components/ProgressLinearColor.vue'
+import TextInput from '../components/TextInput.vue'
+
+Vue.component('progress-linear-color', ProgressLinearColor)
+Vue.component('text-input', TextInput)
